Define initial weather sample Go struct
Almost every field on the struct was returned in a struct with its
value and unit of measure. Thinking about ways of making this Go type
more usable, I think one possiblity is that we could have a Get method
for each field skipping the ".Field.Value" retrieval, returning the
value, and an "ok" indicating whether or not the retrieved field was
actually present.