Handle nullable fields
Re-reading the ClimaCell API docs, I noticed the message that the API
is able to return null values for requested fields on a weather
response if data is not avaliable for that field. Because of this, in
this commit, the FloatValue, IntValue, StringValue, and TimeValue types
are all updated to have pointers for their value fields.

In order to let a user get the value of one of these pointer fields on
a Weather struct, I added GetValue methods to FloatValue, IntValue,
StringValue, and TimeValue, which handles checking for nil pointer
values and dereferencing the values accordingly.

For the ForecastWeather type, I still need to do more experimenting to
see which fields under requested fields are nullable. For example, if
we requested the temperature in a daily forecast, would it be possible
for a daily forecast's high temp value to be null while we know what
time the weather will be the hottest?

This commit also fixes a typo with deserializing 200 responses from
weather data endpoints; getWeatherSamples before was deserializing
those responses into a new []Weather, rather than into the response
interface{} passed in.